Add DebuggerStatement to compiler.compileStatement
Currently, `goja` panics when trying to compile a `debugger` statement.  According to the [spec](http://www.ecma-international.org/ecma-262/6.0/#sec-debugger-statement), a `debugger` statement is a no-op if no debugging facilities are present.  This PR adds an empty `case` statement for `*ast.DebuggerStatement` in `compiler.compileStatement` so that `debugger` statements become no-ops rather than causing the compiler to panic.

Fixes #3